Enhanced results handling for Debrid services, Issue #22 fix, Correctly limit max results option (#25)
* Fixed All-Debrid for season files and corrected maximum results implementation

* Implement max results option in the search

* Issue #22 fix

* Fix for #22 fix

* Correctly limit max results option

* Fix for #22 fix

* Enhanced results handling for Debrid services

---------
