+++ src/helpers/getAvailabilityRD.js
@@ -1,0 +1,13 @@
+export async function getAvailabilityRD(hash, debridApi) {
+    const url = `https://api.real-debrid.com/rest/1.0/torrents/instantAvailability/${hash}`
+	const headers = {
+		Authorization: `Bearer ${debridApi}`,
+	};
+    const response = await fetch(url, { method: "GET", headers });
+    const json = await response.json();
+    const length = json[hash].length;
+    if (length === 0) {
+        return false;
+    }
+    return true;
+}
